[A-4088] Moving resolve team names to service.
