+++ src/client/app/teams/teams.service.ts
@@ -7,15 +7,49 @@ import { Observable } from 'rxjs/Observable';
 import { AsyHttp, HttpOptions } from '../shared/asy-http.service';
 import { PagingOptions } from '../shared/pager.component';
 import { Team, TeamMember } from './teams.class';
+import { User } from '../admin/user.class';
+import { ObservableUtils } from '../shared/observable-utils.class';
+import { ObservableResult } from '../shared/observable-result.class';
 
 @Injectable()
 export class TeamsService {
 
 	cache: any = {};
 
+	teamMap: any = {};
+
 	constructor(
 		private asyHttp: AsyHttp
 	) {
+	}
+
+
+	resolveTeamNames(users: User[]) {
+		if (_.isArray(users)) {
+			// Defensive checking against null teams field
+			users.forEach((user: any) => {
+				user.userModel.teams = user.userModel.teams || [];
+			});
+
+			// Get unique list of team ids to query
+			let teamIds: string[] = _.uniq(_.flatMap(users, (user: any) => user.userModel.teams.map((t: any) => t._id)));
+
+			// Check to see if these ids have been cached
+			let idsToQuery = teamIds.filter((id: string) => !this.teamMap.hasOwnProperty(id));
+
+			// Retrieve data about unknown team ids
+			ObservableUtils.forkJoinSettled(idsToQuery.map((id: string) => this.get(id)))
+				.subscribe(
+					(results: ObservableResult[]) => {
+						results.forEach((result, i) => {
+							if (result.state === 'success') {
+								this.teamMap[result.value._id] = result.value.name;
+							} else {
+								this.teamMap[idsToQuery[i]] = '<missing>';
+							}
+						});
+					});
+		}
 	}
 
 	search(query: any, search: any, paging: PagingOptions, options: any): Observable<Response> {
@@ -38,6 +72,7 @@ export class TeamsService {
 					(results: any) => {
 						if (null != results && _.isArray(results.elements)) {
 							results.elements = results.elements.map((element: any) => new TeamMember().setFromTeamMemberModel(team, element));
+							this.resolveTeamNames(results.elements);
 						}
 						observer.next(results);
 					},
